Rename clustertoolkit to avoid package names with symbols.
