+++ pkg/tools/clustertoolkit/clustertoolkit.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cluster_toolkit
+package clustertoolkit
 
 import (
 	"context"
